requirements.txt
invoke==1.7.1
fabric==2.7.1
requests==2.28.1
pyOpenSSL==22.1.0
PyYAML==6.0
